fix: use int32 instead of int64/long for position IDs
ANE (Apple Neural Engine) operates in FP16 and cannot accurately represent
large int64 values. Position IDs only need integer indexing semantics with
small values, so int32 is sufficient and avoids potential precision loss.

- lm_model_wrapper.py: _generate_position_ids, _generate_causal_mask, and
  LanguageModelWrapper.__init__ all use torch.int32 for position/cache aranges
- glm_ocr_coreml_pipeline.py: position_ids tensors and return cast use int32
- glm_ocr_mtp_model.py: position_id cast uses .to(torch.int32) instead of .long()
- docs/AGENTS.md: update code snippet to reflect int32 usage

+++ src/coremlmodels/glm_ocr/glm_ocr_coreml_pipeline.py
@@ -512,7 +512,7 @@ class GlmOcrCoreMLPipeline:
         # Fallback path when HF reference implementation is unavailable.
         if HFGlmOcrModel is None:
             pos = (
-                torch.arange(seq_len, dtype=torch.long)
+                torch.arange(seq_len, dtype=torch.int32)
                 .view(1, 1, -1)
                 .expand(3, 1, -1)
             )
@@ -530,7 +530,7 @@ class GlmOcrCoreMLPipeline:
             attention_mask=attention_mask_t,
         )
         rope_delta = int(rope_deltas[0, 0].item())
-        return position_ids.to(dtype=torch.long), rope_delta
+        return position_ids.to(dtype=torch.int32), rope_delta
 
     def _compute_chunk_rope_embeddings(
         self,
@@ -557,7 +557,7 @@ class GlmOcrCoreMLPipeline:
     ) -> torch.Tensor:
         """Build 3D position IDs for decode chunk (batch=1)."""
         start = int(cache_position) + int(rope_delta)
-        base = torch.arange(seq_len, dtype=torch.long).view(1, 1, -1) + start
+        base = torch.arange(seq_len, dtype=torch.int32).view(1, 1, -1) + start
         return base.expand(3, 1, -1)
 
     def _prepare_inputs(self, image: Image.Image, prompt: str) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
